Backend/user_management/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-mysql==4.13.0; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt; python_version >= '3.6'
pgsql; python_version >= '3.7'
djangorestframework-simplejwt[crypto]; python_version >= '3.6'
exceptiongroup==1.2.1; python_version <= '3.12'
iniconfig==2.0.0; python_version >= '3.7'
packaging==24.0; python_version >= '3.7'
pika==1.3.2; python_version >= '3.7'
pluggy==1.5.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pytest==8.2.1; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
tomli==2.0.1; python_version <= '3.12'
typing-extensions==4.12.1; python_version <= '3.12'
gunicorn==20.1.0; python_version <= '3.12'
django-environ